INT-29 Fix broken logic
Expected behaviour: If callback verification fails OR invoice state
is not "complete" don't set transcation as complete.

Bug was: If callback verification fails AND invoice state
is not "complete" don't set transcation as complete.

This caused a situation like callback verification passed, invoice state
is "expired" to set the transaction as complete.

This commit fixes it.